Make RAR tests with suggested redirects checking run on .NET Core
