+++ src/Tasks.UnitTests/AssemblyDependency/ResolveAssemblyReferenceTestFixture.cs
@@ -793,10 +793,12 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
             else
             {
                 string gacLocation = null;
+#if FEATURE_GAC
                 if (assemblyName.Version != null)
                 {
                     gacLocation = GlobalAssemblyCache.GetLocation(assemblyName, targetProcessorArchitecture, getRuntimeVersion, targetedRuntimeVersion, fullFusionName, fileExists, null, null, specificVersion /* this value does not matter if we are passing a full fusion name*/);
                 }
+#endif
                 return gacLocation;
             }
         }
